# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Капля API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/groups": {
      "post": {
        "tags": [
          "groups"
        ],
        "summary": "Создать групповую донацию",
        "operationId": "create_group",
        "parameters": [
          {
            "name": "x-max-init-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Max-Init-Data"
            }
          },
          {
            "name": "x-dev-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидна initData",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/groups/my": {
      "get": {
        "tags": [
          "groups"
        ],
        "summary": "Мои предстоящие групповые донации",
        "operationId": "my_groups",
        "parameters": [
          {
            "name": "x-max-init-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Max-Init-Data"
            }
          },
          {
            "name": "x-dev-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Group"
                  },
                  "title": "Response My Groups"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидна initData",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/groups/{code}": {
      "get": {
        "tags": [
          "groups"
        ],
        "summary": "Группа по коду из ссылки grp_<code>",
        "operationId": "get_group",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "x-max-init-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Max-Init-Data"
            }
          },
          {
            "name": "x-dev-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидна initData",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/groups/{code}/join": {
      "post": {
        "tags": [
          "groups"
        ],
        "summary": "Присоединиться к группе",
        "operationId": "join_group",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "x-max-init-data",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Max-Init-Data"
            }
          },
          {
            "name": "x-dev-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидна initData",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Конфликт",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DonationType": {
        "type": "string",
        "enum": [
          "whole_blood",
          "plasma"
        ],
        "title": "DonationType",
        "description": "Вид донации: цельная кровь или плазма."
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "fields": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "fields"
        ],
        "title": "ErrorBody"
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "unauthorized",
          "not_found",
          "validation_error",
          "slot_not_found",
          "slot_taken",
          "active_exists",
          "interval_not_passed",
          "personal_data_incomplete",
          "appointment_not_found",
          "appointment_not_active",
          "invite_not_found",
          "demo_disabled",
          "donation_not_found",
          "group_not_found",
          "group_closed",
          "bot_send_failed",
          "internal_error"
        ],
        "title": "ErrorCode"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "Group": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "center": {
            "$ref": "#/components/schemas/GroupCenter"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "donation_type": {
            "$ref": "#/components/schemas/DonationType"
          },
          "owner_name": {
            "type": "string",
            "title": "Owner Name"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/GroupMember"
            },
            "type": "array",
            "title": "Members"
          },
          "members_count": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Members Count"
          },
          "is_member": {
            "type": "boolean",
            "title": "Is Member"
          },
          "is_owner": {
            "type": "boolean",
            "title": "Is Owner"
          },
          "is_booked": {
            "type": "boolean",
            "title": "Is Booked",
            "description": "Текущий пользователь уже записан"
          },
          "is_past": {
            "type": "boolean",
            "title": "Is Past"
          },
          "free_slots": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Free Slots",
            "description": "Свободных слотов в центре на дату"
          },
          "link": {
            "type": "string",
            "title": "Link",
            "description": "https://max.ru/<бот>?startapp=grp_<code>"
          },
          "share_text": {
            "type": "string",
            "title": "Share Text",
            "description": "Текст для shareMaxContent"
          }
        },
        "type": "object",
        "required": [
          "code",
          "center",
          "date",
          "donation_type",
          "owner_name",
          "members",
          "members_count",
          "is_member",
          "is_owner",
          "is_booked",
          "is_past",
          "free_slots",
          "link",
          "share_text"
        ],
        "title": "Group"
      },
      "GroupCenter": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "region_id": {
            "type": "integer",
            "title": "Region Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "address",
          "region_id"
        ],
        "title": "GroupCenter"
      },
      "GroupCreate": {
        "properties": {
          "center_id": {
            "type": "integer",
            "title": "Center Id"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date",
            "description": "Локальная дата центра"
          },
          "donation_type": {
            "$ref": "#/components/schemas/DonationType"
          }
        },
        "type": "object",
        "required": [
          "center_id",
          "date",
          "donation_type"
        ],
        "title": "GroupCreate"
      },
      "GroupMember": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Имя и первая буква фамилии: «Иван П.»"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "is_owner": {
            "type": "boolean",
            "title": "Is Owner"
          },
          "is_booked": {
            "type": "boolean",
            "title": "Is Booked",
            "description": "Есть запись в этот центр на эту дату"
          }
        },
        "type": "object",
        "required": [
          "name",
          "photo_url",
          "is_owner",
          "is_booked"
        ],
        "title": "GroupMember"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}